After much frustration, Buoys are now called Sites! Hurrah!
Also, Contacts are now serialised once each and are linked to their respective sites using an ID number system. There are several tests for all of these, and several tests have been amended as a result of the changes made.

Switched serialiser to a more human-readable one, so it is far more predictable with what comes out, and makes our tests more authentic because we can actually produce XML to test for! Result!

One note: the selected site in the list box within the session view isn't binding properly at this stage. Will look into it right now, along with fixing the disabled/enabled states for the contact editing within the session view.

Signed-off-by: Steven McTainsh <smctainsh@gmail.com>